# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoStake Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/me/activity:
    get:
      operationId: getMyActivity
      tags: [users]
      summary: Get recent activity feed for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── Wallet ──────────────────────────────────────────────────────────────────
...
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request

  # ─── Transactions ─────────────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions for current user
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, bet, win, loss]
        - name: currency
          in: query
          schema:
            type: string
            enum: [ETH, BTC, SOL, USDC]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found

  # ─── KYC / Identity Verification ─────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, email, createdAt, kycStatus, isAdmin]
      properties:
        id:          { type: integer }
        clerkId:     { type: string }
        username:    { type: string }
        email:       { type: string }
        createdAt:   { type: string }
        kycStatus:   { type: string, enum: [none, pending, approved, rejected] }
        isAdmin:     { type: boolean }
        avatarUrl:   { type: ["string", "null"] }
    UserUpdate:
      type: object
      properties:
        username:  { type: string }
        avatarUrl: { type: string }
...
    ActivityItem:
      type: object
      required: [id, type, description, amount, currency, createdAt]
      properties:
        id:          { type: integer }
        type:        { type: string }
        description: { type: string }
        amount:      { type: string }
        currency:    { type: string }
        createdAt:   { type: string }

    # ─── Wallet ──────────────────────────────────────────────────────────────
...
    WithdrawalRequest:
      type: object
      required: [currency, amount, toAddress]
      properties:
        currency:  { type: string, enum: [ETH, BTC, SOL, USDC] }
        amount:    { type: string }
        toAddress: { type: string }

    # ─── Transactions ────────────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, userId, type, currency, amount, status, createdAt]
      properties:
        id:          { type: integer }
        userId:      { type: integer }
        type:        { type: string, enum: [deposit, withdrawal, bet, win, loss] }
        currency:    { type: string }
        amount:      { type: string }
        status:      { type: string, enum: [pending, confirmed, failed] }
        txHash:      { type: ["string", "null"] }
        toAddress:   { type: ["string", "null"] }
        gameType:    { type: ["string", "null"] }
        createdAt:   { type: string }
    TransactionPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }

    # ─── KYC ─────────────────────────────────────────────────────────────────
...