# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CryptoStake Casino API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  amount: string;
  currency: string;
  createdAt: string;
}