# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CryptoStake Casino API
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from './userKycStatus';

export interface User {
  id: number;
  clerkId: string;
  username: string;
  email: string;
  createdAt: string;
  kycStatus: UserKycStatus;
  isAdmin: boolean;
  /** @nullable */
  avatarUrl?: string | null;
}